Reorganize repo after JS enlightenment.
