+++ package.json
@@ -8,7 +8,6 @@
   "unpkg": "dist/IndralabVue.umd.min.js",
   "jsdelivr": "dist/IndralabVue.umd.min.js",
   "scripts": {
-    "serve": "vue-cli-service serve",
     "build": "vue-cli-service build --target lib --name IndralabVue src/index.js",
     "lint": "vue-cli-service lint",
     "watch": "npm run build -- --watch",
@@ -17,16 +16,17 @@
   },
   "dependencies": {
     "core-js": "^3.6.5",
-    "vue": "^2.6.12"
+    "vue": "^3.0.11",
+    "vue-router": "^4.0.8"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.5.4",
-    "@vue/cli-plugin-eslint": "^4.5.4",
-    "@vue/cli-service": "^4.5.11",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.19.0",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
+    "babel-eslint": "^10.0.3",
+    "eslint": "^5.16.0",
     "eslint-plugin-vue": "^5.0.0",
-    "minimist": ">=1.2.3",
     "vue-template-compiler": "^2.6.12"
   },
   "eslintConfig": {
