--- src/components/AgentPair/index.js
@@ -1,5 +1,0 @@
-import ilComponent from './AgentPair'
-
-export default Vue => {
-    Vue.component(ilComponent.name, ilComponent);
-}
