--- src/components/AgentSelect/index.js
@@ -1,6 +1,0 @@
-import ilComponent from './AgentSelect'
-
-export default Vue => {
-    Vue.component(ilComponent.name, ilComponent);
-}
-
