--- src/components/CurationRow/index.js
@@ -1,6 +1,0 @@
-import ilComponent from './CurationRow'
-
-export default Vue => {
-    Vue.component(ilComponent.name, ilComponent);
-}
-
