--- src/components/Evidence/index.js
@@ -1,6 +1,0 @@
-import ilComponent from './Evidence'
-
-export default Vue => {
-    Vue.component(ilComponent.name, ilComponent);
-}
-
