--- src/components/MeshSelect/index.js
@@ -1,6 +1,0 @@
-import ilComponent from './MeshSelect'
-
-export default Vue => {
-    Vue.component(ilComponent.name, ilComponent);
-}
-
