--- src/components/PaperSelect/index.js
@@ -1,6 +1,0 @@
-import ilComponent from './PaperSelect'
-
-export default Vue => {
-    Vue.component(ilComponent.name, ilComponent);
-}
-
