--- src/components/RefLink/index.js
@@ -1,6 +1,0 @@
-import ilComponent from './RefLink'
-
-export default Vue => {
-    Vue.component(ilComponent.name, ilComponent);
-}
-
