--- src/components/Relation/index.js
@@ -1,5 +1,0 @@
-import ilComponent from './Relation'
-
-export default Vue => {
-    Vue.component(ilComponent.name, ilComponent);
-}
