--- src/components/RelationSearch/index.js
@@ -1,5 +1,0 @@
-import ilComponent from './RelationSearch'
-
-export default Vue => {
-    Vue.component(ilComponent.name, ilComponent);
-}
