--- src/components/SourceDisplay/index.js
@@ -1,5 +1,0 @@
-import ilComponent from './SourceDisplay'
-
-export default Vue => {
-    Vue.component(ilComponent.name, ilComponent);
-}
