--- src/components/Statement/index.js
@@ -1,5 +1,0 @@
-import ilComponent from './Statement'
-
-export default Vue => {
-    Vue.component(ilComponent.name, ilComponent);
-}
