--- src/components/TypeSelect/index.js
@@ -1,6 +1,0 @@
-import ilComponent from './TypeSelect'
-
-export default Vue => {
-    Vue.component(ilComponent.name, ilComponent);
-}
-
