+++ src/components/constraints/index.js
@@ -1,0 +1,4 @@
+export {default as AgentSelect} from './AgentSelect'
+export {default as TypeSelect} from './TypeSelect'
+export {default as MeshSelect} from './MeshSelect'
+export {default as PaperSelect} from './PaperSelect'
