+++ src/components/result_display/AgentPair.vue
@@ -44,7 +44,7 @@
 </template>
 
 <script>
-  import piecemeal_mixin from '../piecemeal_mixin'
+  import piecemeal_mixin from '../mixins/piecemeal_mixin'
 
   export default {
     name: "AgentPair",
