+++ src/components/result_display/CurationRow.vue
