+++ src/components/result_display/index.js
@@ -1,0 +1,7 @@
+export { default as AgentPair } from './AgentPair'
+export { default as Relation } from './Relation'
+export { default as Statement } from './Statement'
+export { default as Evidence } from './Evidence'
+export { default as CurationRow } from './CurationRow'
+export { default as RefLink } from './RefLink'
+export { default as SourceDisplay } from './SourceDisplay'
